ov025: promote daDkk_c into a single C++ TU
Absorbs all eight one-function shards at 0x021118c8..0x02111cf4 into
src/actors/daDkk_c.cpp. tubuild verify reports 8/8 MATCH, objisolate clean,
reloc-destinations clean, and all eight sections in ROM-ascending emission
order; tubuild linkcheck reports SCRATCH-LINK-VERIFIED with the range
byte-identical, 106/106 modules exact, and a full ROM whose sha256 matches the
stock build. dsd check symbols --fail still fails on the same nine pre-existing
errors the baseline control has -- 0 new, 0 resolved.

WHOLE promotion, destructor pair included. The cartridge orders D1 (0x021118c8)
below D0 (0x02111928), which is the direction an in-class destructor
reproduces, so include/daDkk_c.h moves ~daDkk_c() into the class body. That is
a byte requirement rather than a style choice: the out-of-line form emits
D2, D0, D1 plus a homeless D2 the ROM does not carry. It also moves the class's
key function down to InitResources, which this TU defines. Its own base class
daDsnBase_c is the opposite case -- ov091 orders that one D0-below-D1, which is
why it promoted 9-of-11 while this leaf promotes whole.

compiler_only_output is MEASURED, not estimated: 14 rows, this class having had
no sibling oracle in the queue. Thirteen are RTTI and vtable records -- the
__si_class_type_info chain in ov025/relocs.txt gives five ancestors, so
2 x 6 + 1 = 13 -- and the fourteenth is _ZN7Vector3D1Ev, emitted because the TU
takes Vector3 locals, not because any member has that type. Every one of the
thirteen has a configured ROM home, in arm9, ov002 or ov025, so all are
deadstrip-data and externalized_output is empty. Unlike daDsnBase_c there is no
@class$ anonymous-struct destructor: that appears only when the source declares
a file-local anonymous struct to hold the Vector3 locals.

The _ZTV7daDkk_c extent is read from relocations, not symbols.txt spacing:
every word 0x0211384c..0x021138cc is a relocated entry and 0x021138d0 already
belongs to _ZTI14daObjDpBrock_c, giving 2 preamble words + 32 slots = 0x88.
Three ambiguous data_ov025_* rows sit inside the table and would have given a
false 0x58.

Reconciliations, all recorded in the manifest notes: decl_common.h's char*
spellings replace the shards' void* ones for ten address-named symbols;
func_0201267c's two contradictory return types (int in one shard, void in the
other, undetected because tubuild's conflict detector does not compare return
types) settle on void, which no call site here can observe; the two RAW-flagged
shards become plain file-scope extern "C" definitions; and both destructor
forcing scaffolds are dropped, owning the key function being enough.
