+++ config/arm9/overlays/ov025/delinks.txt
@@ -44,37 +44,9 @@ src/d_a_dgr.c:
     complete
     .text start:0x02111898 end:0x021118c8
 
-src/_ZN7daDkk_cD1Ev.cpp:
+src/actors/daDkk_c.cpp:
     complete
-    .text start:0x021118c8 end:0x02111928
-
-src/_ZN7daDkk_cD0Ev.cpp:
-    complete
-    .text start:0x02111928 end:0x0211199c
-
-src/_ZN7daDkk_c16OnAimedAtWithEggEv.cpp:
-    complete
-    .text start:0x0211199c end:0x021119a4
-
-src/func_ov025_021119a4.cpp:
-    complete
-    .text start:0x021119a4 end:0x021119f4
-
-src/func_ov025_021119f4.c:
-    complete
-    .text start:0x021119f4 end:0x02111a84
-
-src/func_ov025_02111a84.cpp:
-    complete
-    .text start:0x02111a84 end:0x02111b64
-
-src/_ZN7daDkk_c8BehaviorEv.cpp:
-    complete
-    .text start:0x02111b64 end:0x02111c24
-
-src/_ZN7daDkk_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x02111c24 end:0x02111cf4
+    .text start:0x021118c8 end:0x02111cf4
 
 src/d_a_dkk.c:
     complete
