+++ config/converted-backslide-exceptions.jsonl
@@ -329,3 +329,6 @@
 {"path": "src/_ZN8StarDoorD0Ev.cpp", "reason": "ov100/daStarGate_c promotion consolidates legacy StarDoor and particle callback sources into the exact enrolled translation unit; member identities replace retired one-function paths."}
 {"path": "src/_ZN8StarDoorD1Ev.cpp", "reason": "ov100/daStarGate_c promotion consolidates legacy StarDoor and particle callback sources into the exact enrolled translation unit; member identities replace retired one-function paths."}
 {"path": "src/__sinit_ov100_02147a70.c", "reason": "ov100/daStarGate_c promotion consolidates legacy StarDoor and particle callback sources into the exact enrolled translation unit; member identities replace retired one-function paths."}
+{"path": "src/_ZN7daDkk_c16OnAimedAtWithEggEv.cpp", "reason": "ov025/daDkk_c promoted: 8 one-function shards absorbed into src/actors/daDkk_c.cpp"}
+{"path": "src/_ZN7daDkk_cD0Ev.cpp", "reason": "ov025/daDkk_c promoted: 8 one-function shards absorbed into src/actors/daDkk_c.cpp"}
+{"path": "src/_ZN7daDkk_cD1Ev.cpp", "reason": "ov025/daDkk_c promoted: 8 one-function shards absorbed into src/actors/daDkk_c.cpp"}
