+++ config/converted-baseline.json
@@ -1544,9 +1544,6 @@
     "src/_ZN7daDgr_c6RenderEv.cpp",
     "src/_ZN7daDgr_cD0Ev.cpp",
     "src/_ZN7daDgr_cD1Ev.cpp",
-    "src/_ZN7daDkk_c16OnAimedAtWithEggEv.cpp",
-    "src/_ZN7daDkk_cD0Ev.cpp",
-    "src/_ZN7daDkk_cD1Ev.cpp",
     "src/_ZN7daKrb_c16OnPendingDestroyEv.cpp",
     "src/_ZN7daKrb_cD0Ev.cpp",
     "src/_ZN7daKrb_cD1Ev.cpp",
@@ -2150,6 +2147,9 @@
     "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_c24OnHitByCannonBlastedCharEv",
     "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_cD0Ev",
     "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_cD1Ev",
+    "src/actors/daDkk_c.cpp#_ZN7daDkk_c16OnAimedAtWithEggEv",
+    "src/actors/daDkk_c.cpp#_ZN7daDkk_cD0Ev",
+    "src/actors/daDkk_c.cpp#_ZN7daDkk_cD1Ev",
     "src/actors/daDsnBase_c.cpp#_ZN11daDsnBase_c16CleanupResourcesEv",
     "src/actors/daObjCtMecha03_c.cpp#_ZN16daObjCtMecha03_c6RenderEv",
     "src/actors/daObjCtMecha03_c.cpp#_ZN16daObjCtMecha03_cD0Ev",
